requirements.txt
# Core data processing dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
scikit-learn>=1.0.0,<2.0.0
pyyaml>=6.0

# Progress bars for long-running operations
tqdm>=4.62.0

# Data cleaning (separate repository - install manually)
# git clone https://github.com/mdparkes/datacleaner/ && pip install ./datacleaner
# datacleaner